requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-xml==2.0.0
idna==3.3
numpy==1.23.2
pandas==1.4.4
python-dateutil==2.8.2
pytz==2022.2.1
regex-engine==1.1.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
